% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnk_proj_byname.R
\name{mnk_proj_byname}
\alias{mnk_proj_byname}
\title{mnk_proj_byname
Get information on a specific project selected by  a string contained in the project name}
\usage{
mnk_proj_byname(query)
}
\arguments{
\item{query}{A string that is contained in the project name}
}
\value{
A data frame with all the projects that contain the string with some details of those projects
}
\description{
Get information on a specific project selected by  a string contained in the project name
}
\examples{
\dontrun{
mnk_obs <- mnk_proj_byname(query="Biomarato 2025")
mnk_obs_id(m_obs$id[1])
}
}
