% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnk_proj_info.R
\name{mnk_proj_info}
\alias{mnk_proj_info}
\title{mnk_proj_info
Download complete info from a Minka project}
\usage{
mnk_proj_info(project_id = NULL, grpid = NULL)
}
\arguments{
\item{project_id}{A single id for a Minka project.}

\item{grpid}{Name of the group as an Minka slug or group ID.
returns the raw output of parsed JSON for that project. Otherwise just some basic information is returned as a list.}
}
\value{
A list containing project details similar to what you can find on a project's page.
}
\description{
This function can be used to get either information from a project
by complete project name or ID.
}
\details{
Before using this function, you need to know the \code{project_id} or the project slug name.
If the ID or the slug name is unknown, you can first use the function \code{\link{mnk_proj_byname}}
to find it based on the project's name.
A Minka slug is usually the project name as a single string with words separated by hyphens.
For instance, the project "biomarato-2022-girona" has a slug of "world-oceans-week-2022",
which you can find by searching projects on Minka and looking at the \href{https://minka-sdg.org/}{project's page's URL}.
}
\examples{
\dontrun{
 mnk_proj_byname("Biomarato Barcelona 2025")
 mnk_minka_proj_info("420")
}
}
