% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnk_proj_obs.R
\name{mnk_proj_obs}
\alias{mnk_proj_obs}
\title{mnk_proj_obs
Download observations from a Minka project}
\usage{
mnk_proj_obs(project_id, year, month = NULL)
}
\arguments{
\item{project_id}{A single id for a Minka project.}

\item{year}{Return all observations of project only in that year (can only be
one year, not a range of years).}

\item{month}{A number from 1 to 12 to download only that month.
If it is NULL, the entire year will be downloaded.}
}
\value{
A data frame with all the year/month records with full details.
}
\description{
This function downloads observation data for a Minka project,in a given year y,
optionally, a specific month. It automatically manages if you download it
must be done month by month or for a single month.
}
\details{
Before using this function, you need to know the \code{project_id}.
If the ID is unknown, you can first use the function \code{\link{mnk_proj_byname}}
to find it based on the project's name.
}
\examples{
\dontrun{
  #First of all it is necessary to obtain the project_id of the project.
   projects <- mnk_proj_byname(query="Biomarato 2025")

  #Select the id_project
  obs_june_2025 <- mnk_proj_obs( projects$project_id[1], year=2025, month=6)
  obs_2025 <- mnk_proj_obs( projects$project_id[1], year=2025, month=NULL)
}
}
