% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnk_user_proj.R
\name{mnk_user_proj}
\alias{mnk_user_proj}
\title{mnk_user_proj}
\usage{
mnk_user_proj(id_user)
}
\arguments{
\item{id_user}{(numeric): The user ID}
}
\value{
A tibble where each row is a project and the columns
contain relevant information about the project.
Returns an empty tibble if there are no projects or if an error occurs.
}
\description{
Retrieves the projects to which a user is subscribed from the Minka-SDG API
    and returns them as a tibble.
}
\examples{
\dontrun{
# projects_for_user_6 <- mnk_user_proj(6)
# print(projects_for_user_6)
}

}
