% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnk_places_byname.R
\name{mnk_places_byname}
\alias{mnk_places_byname}
\title{mnk_place_byname
Get information about Minka place by name}
\usage{
mnk_places_byname(query)
}
\arguments{
\item{query}{A single id for a Minka observation record}
}
\value{
A tibble with all details on a given record
}
\description{
Get information about Minka place by name
}
\examples{
\dontrun{
sant_feliu <- mnk_places_byname(query="Sant Feliu")
}
}
